This commit is of the libraries that I used I created/used to make this project work correctly. The TouchScreen and TFTScreen libraries are *heavily* modified, rewritten, cut down and optimised from the original mess of libraries that was provided on the touch screen shield product page from DX.com. The keyboard library came with the Arduino software and the Adafruit-GFX library was provided from the product page but was originally created for an Adafruit touch screen shield. The rotary encoder library was pulled from https://github.com/mathertel/RotaryEncoder.
These libraries appear to all be the ones used when making this project, but I don't have sequential versions of the libraries that I modified and rewrote so they may not be compatible with all of the Arduino sketches. I'm also hesitant to test them because there's no way to revert the Arduino back to the version which the Photoshop Controller is currently running, works and I use fairly often.